Add gamepad axis inversion toggles
git-svn-id: https://svn.eduke32.com/eduke32@7986 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/mact/include/_control.h
@@ -171,8 +171,8 @@ typedef struct
 typedef struct
 {
     int32_t analog;
-    int32_t digital;
-    int16_t digitalClearedN, digitalClearedP;
+    int8_t digital;
+    int8_t digitalClearedN, digitalClearedP;
 } controlaxistype;
